+++ cmd/traffic/cmd/manager/internal/mutator/agent_injector.go
@@ -3,6 +3,7 @@ package mutator
 import (
 	"context"
 	"fmt"
+	"sort"
 	"strconv"
 	"strings"
 
@@ -131,6 +132,11 @@ func agentInjector(ctx context.Context, req *admission.AdmissionRequest) ([]patc
 	} else {
 		patches = hidePorts(&pod, appContainer, servicePort.TargetPort.StrVal, patches)
 	}
+	tpEnv := make(map[string]string)
+	if env.APIPort != 0 {
+		tpEnv["TELEPRESENCE_API_PORT"] = strconv.Itoa(int(env.APIPort))
+	}
+	patches = addTPEnv(&pod, appContainer, tpEnv, patches)
 	patches, err = addAgentContainer(ctx, svc, &pod, servicePort, appContainer, &appPort, setGID, podName, podNamespace, patches)
 	if err != nil {
 		return nil, err
@@ -274,11 +280,53 @@ func addAgentContainer(
 			appContainer,
 			containerPort,
 			int(appPort.ContainerPort),
+			int(env.APIPort),
 			env.ManagerNamespace,
 			setGID,
 		)})
 
 	return patches, nil
+}
+
+// addTPEnv adds telepresence specific environment variables to the app container
+func addTPEnv(pod *corev1.Pod, cn *corev1.Container, env map[string]string, patches []patchOperation) []patchOperation {
+	if len(env) == 0 {
+		return patches
+	}
+	cns := pod.Spec.Containers
+	var containerPath string
+	for i := range cns {
+		if &cns[i] == cn {
+			containerPath = fmt.Sprintf("/spec/containers/%d", i)
+			break
+		}
+	}
+	keys := make([]string, len(env))
+	i := 0
+	for k := range env {
+		keys[i] = k
+		i++
+	}
+	sort.Strings(keys)
+	if cn.Env == nil {
+		patches = append(patches, patchOperation{
+			Op:    "replace",
+			Path:  fmt.Sprintf("%s/%s", containerPath, "env"),
+			Value: []corev1.EnvVar{},
+		})
+	}
+	for _, k := range keys {
+		patches = append(patches, patchOperation{
+			Op:   "add",
+			Path: fmt.Sprintf("%s/%s", containerPath, "env/-"),
+			Value: corev1.EnvVar{
+				Name:      k,
+				Value:     env[k],
+				ValueFrom: nil,
+			},
+		})
+	}
+	return patches
 }
 
 // hidePorts  will replace the symbolic name of a container port with a generated name. It will perform