+++ pkg/client/connector/userd_trafficmgr/install.go
@@ -30,7 +30,7 @@ type installer struct {
 
 type Installer interface {
 	userd_k8s.ResourceFinder
-	EnsureAgent(c context.Context, namespace, name, svcName, portNameOrNumber, agentImageName string) (string, string, error)
+	EnsureAgent(c context.Context, namespace, name, svcName, portNameOrNumber, agentImageName string, telepresenceAPIPort uint16) (string, string, error)
 	EnsureManager(c context.Context) error
 	RemoveManagerAndAgents(c context.Context, agentsOnly bool, agents []*manager.AgentInfo) error
 }
@@ -263,7 +263,8 @@ func (ki *installer) getSvcForInjectedPod(
 // is installed alongside the proper workload. In doing that, it also ensures that
 // the workload is referenced by a service. Lastly, it returns the service UID
 // associated with the workload since this is where that correlation is made.
-func (ki *installer) EnsureAgent(c context.Context, namespace, name, svcName, portNameOrNumber, agentImageName string) (string, string, error) {
+func (ki *installer) EnsureAgent(c context.Context,
+	namespace, name, svcName, portNameOrNumber, agentImageName string, telepresenceAPIPort uint16) (string, string, error) {
 	obj, err := ki.FindWorkload(c, namespace, name)
 	if err != nil {
 		return "", "", err
@@ -309,6 +310,7 @@ already exist for this service`, kind, obj.GetName())
 	}
 
 	update := true
+	updateSvc := false
 	switch {
 	case agentContainer == nil:
 		dlog.Infof(c, "no agent found for %s %s.%s", kind, name, namespace)
@@ -319,7 +321,7 @@ already exist for this service`, kind, obj.GetName())
 		if err != nil {
 			return "", "", err
 		}
-		obj, svc, err = addAgentToWorkload(c, portNameOrNumber, agentImageName, ki.GetManagerNamespace(), obj, matchingSvc)
+		obj, svc, updateSvc, err = addAgentToWorkload(c, portNameOrNumber, agentImageName, ki.GetManagerNamespace(), telepresenceAPIPort, obj, matchingSvc)
 		if err != nil {
 			return "", "", err
 		}
@@ -351,7 +353,7 @@ already exist for this service`, kind, obj.GetName())
 		if err := ki.Client().Update(c, obj, obj); err != nil {
 			return "", "", err
 		}
-		if svc != nil {
+		if updateSvc {
 			if err := ki.Client().Update(c, svc, svc); err != nil {
 				return "", "", err
 			}
@@ -590,21 +592,23 @@ func addAgentToWorkload(
 	portNameOrNumber string,
 	agentImageName string,
 	trafficManagerNamespace string,
+	telepresenceAPIPort uint16,
 	object kates.Object, matchingService *kates.Service,
 ) (
 	kates.Object,
 	*kates.Service,
+	bool,
 	error,
 ) {
 	podTemplate, err := install.GetPodTemplateFromObject(object)
 	if err != nil {
-		return nil, nil, err
+		return nil, nil, false, err
 	}
 
 	cns := podTemplate.Spec.Containers
 	servicePort, container, containerPortIndex, err := install.FindMatchingPort(cns, portNameOrNumber, matchingService)
 	if err != nil {
-		return nil, nil, install.ObjErrorf(object, err.Error())
+		return nil, nil, false, install.ObjErrorf(object, err.Error())
 	}
 	dlog.Debugf(c, "using service %q port %q when intercepting %s %q",
 		matchingService.Name,
@@ -657,7 +661,7 @@ func addAgentToWorkload(
 		}
 	}
 	if containerPort.Number == 0 {
-		return nil, nil, install.ObjErrorf(object, "unable to add: the container port cannot be determined")
+		return nil, nil, false, install.ObjErrorf(object, "unable to add: the container port cannot be determined")
 	}
 	if containerPort.Name == "" {
 		containerPort.Name = fmt.Sprintf("tx-%d", containerPort.Number)
@@ -671,6 +675,14 @@ func addAgentToWorkload(
 			AppPortProto:  containerPort.Protocol,
 			AppPortNumber: containerPort.Number,
 			ImageName:     agentImageName,
+		}
+	}
+
+	var addTPEnvAction *addTPEnvironmentAction
+	if telepresenceAPIPort != 0 {
+		addTPEnvAction = &addTPEnvironmentAction{
+			ContainerName: container.Name,
+			Env:           map[string]string{"TELEPRESENCE_API_PORT": strconv.Itoa(int(telepresenceAPIPort))},
 		}
 	}
 
@@ -688,8 +700,10 @@ func addAgentToWorkload(
 			ContainerPortName:       containerPort.Name,
 			ContainerPortProto:      containerPort.Protocol,
 			ContainerPortNumber:     containerPort.Number,
+			APIPortNumber:           telepresenceAPIPort,
 			ImageName:               agentImageName,
 		},
+		AddTPEnvironmentAction: addTPEnvAction,
 	}
 	// Depending on whether the Service refers to the port by name or by number, we either need
 	// to patch the names in the deployment, or the number in the service.
@@ -733,7 +747,7 @@ func addAgentToWorkload(
 
 	// Apply the actions on the workload.
 	if err = workloadMod.Do(object); err != nil {
-		return nil, nil, err
+		return nil, nil, false, err
 	}
 	annotations := object.GetAnnotations()
 	if object.GetAnnotations() == nil {
@@ -741,29 +755,29 @@ func addAgentToWorkload(
 	}
 	annotations[annTelepresenceActions], err = workloadMod.MarshalAnnotation()
 	if err != nil {
-		return nil, nil, err
+		return nil, nil, false, err
 	}
 	object.SetAnnotations(annotations)
 	explainDo(c, workloadMod, object)
 
 	// Apply the actions on the Service.
+	updateService := false
 	if serviceMod != nil {
 		if err = serviceMod.Do(matchingService); err != nil {
-			return nil, nil, err
+			return nil, nil, false, err
 		}
 		if matchingService.Annotations == nil {
 			matchingService.Annotations = make(map[string]string)
 		}
 		matchingService.Annotations[annTelepresenceActions], err = serviceMod.MarshalAnnotation()
 		if err != nil {
-			return nil, nil, err
+			return nil, nil, false, err
 		}
 		explainDo(c, serviceMod, matchingService)
-	} else {
-		matchingService = nil
+		updateService = true
 	}
 
-	return object, matchingService, nil
+	return object, matchingService, updateService, nil
 }
 
 func (ki *installer) EnsureManager(c context.Context) error {