+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 2.4.8 (TBD)
+
+- Feature: A RESTful service was added to Telepresence, both locally to the client and to the `traffic-agent` to help determine if messages with a set of headers should be
+  consumed or not from a message queue where the intercept headers are added to the messages.
+
 ### 2.4.7 (November 24, 2021)
 
 - Feature: The agent injector now supports a new annotation, `telepresence.getambassador.io/inject-service-name`, that can be used to set the name of the service to be intercepted.